Completado a requisição de Compra e criado o auxiliar ItemCompra para salvar os produtos

+++ api-security/src/main/java/api_security/api_security/user/dto/ProdutoDto.java
@@ -1,5 +1,5 @@
 package api_security.api_security.user.dto;
 
-public record ProdutoDto(Integer produtoId, Integer quantidade) {
+public record ProdutoDto(Integer produtoId, Double quantidade) {
     
 }
